Furm/requirements.txt
django==4.1.1
django-environ==0.9.0
djangorestframework==3.14.0
django-widget-tweaks==1.4.12
